moved 0.5 out of feature flag, removed deprecated 0.4 code and features
