+++ drone/server/server.go
@@ -86,7 +86,6 @@ func start(c *cli.Context) error {
 		middleware.Cache(),
 		middleware.Store(),
 		middleware.Remote(),
-		middleware.Engine(),
 	)
 
 	if c.String("server-cert") != "" {
@@ -109,7 +108,7 @@ var agreement = `
 
 
 You are attempting to use the unstable channel. This build is experimental and
-has known bugs and compatibility issues, and is not intended for general use.
+has known bugs and compatibility issues. It is not intended for general use.
 
 Please consider using the latest stable release instead:
 
@@ -119,8 +118,8 @@ If you are attempting to build from source please use the latest stable tag:
 
 		v0.4.2
 
-If you are interested in testing this experimental build and assisting with
-development you will need to set the following environment variables to proceed:
+If you are interested in testing this experimental build AND assisting with
+development you may proceed by setting the following environment:
 
 		I_UNDERSTAND_I_AM_USING_AN_UNSTABLE_VERSION=true
 		I_AGREE_TO_FIX_BUGS_AND_NOT_FILE_BUGS=true
