+++ store/store.go
@@ -125,21 +125,6 @@ type Store interface {
 
 	// WriteLog writes the job logs to the datastore.
 	WriteLog(*model.Job, io.Reader) error
-
-	// GetNode gets a build node from the datastore.
-	GetNode(id int64) (*model.Node, error)
-
-	// GetNodeList gets a build node list from the datastore.
-	GetNodeList() ([]*model.Node, error)
-
-	// CreateNode add a new build node to the datastore.
-	CreateNode(*model.Node) error
-
-	// UpdateNode updates a build node in the datastore.
-	UpdateNode(*model.Node) error
-
-	// DeleteNode removes a build node from the datastore.
-	DeleteNode(*model.Node) error
 }
 
 // GetUser gets a user by unique ID.
@@ -342,24 +327,4 @@ func ReadLog(c context.Context, job *model.Job) (io.ReadCloser, error) {
 
 func WriteLog(c context.Context, job *model.Job, r io.Reader) error {
 	return FromContext(c).WriteLog(job, r)
-}
-
-func GetNode(c context.Context, id int64) (*model.Node, error) {
-	return FromContext(c).GetNode(id)
-}
-
-func GetNodeList(c context.Context) ([]*model.Node, error) {
-	return FromContext(c).GetNodeList()
-}
-
-func CreateNode(c context.Context, node *model.Node) error {
-	return FromContext(c).CreateNode(node)
-}
-
-func UpdateNode(c context.Context, node *model.Node) error {
-	return FromContext(c).UpdateNode(node)
-}
-
-func DeleteNode(c context.Context, node *model.Node) error {
-	return FromContext(c).DeleteNode(node)
 }
