+++ web/pages.go
@@ -30,7 +30,7 @@ func ShowIndex(c *gin.Context) {
 	}
 
 	// filter to only show the currently active ones
-	activeRepos, err := store.GetRepoListOf(c,repos)
+	activeRepos, err := store.GetRepoListOf(c, repos)
 	if err != nil {
 		c.String(400, err.Error())
 		return
@@ -80,26 +80,6 @@ func ShowUser(c *gin.Context) {
 	c.HTML(200, "user.html", gin.H{
 		"User": user,
 		"Csrf": token,
-	})
-}
-
-func ShowUsers(c *gin.Context) {
-	user := session.User(c)
-	if !user.Admin {
-		c.AbortWithStatus(http.StatusForbidden)
-		return
-	}
-	users, _ := store.GetUserList(c)
-
-	token, _ := token.New(
-		token.CsrfToken,
-		user.Login,
-	).Sign(user.Hash)
-
-	c.HTML(200, "users.html", gin.H{
-		"User":  user,
-		"Users": users,
-		"Csrf":  token,
 	})
 }
 
@@ -226,11 +206,4 @@ func ShowBuild(c *gin.Context) {
 		"Job":   job,
 		"Csrf":  csrf,
 	})
-}
-
-func ShowNodes(c *gin.Context) {
-	user := session.User(c)
-	nodes, _ := store.GetNodeList(c)
-	token, _ := token.New(token.CsrfToken, user.Login).Sign(user.Hash)
-	c.HTML(http.StatusOK, "nodes.html", gin.H{"User": user, "Nodes": nodes, "Csrf": token})
 }
